pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
 
[project]
name = "azure-maestro"
version = "0.1.0"
description = "A CLI tool for orchestrating Azure Container Apps using Italian musical terminology"
authors = [{ name = "Dimitrios Souflis", email = "dsouflis@gmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pyyaml>=6.0",
    "pydantic==2.11.7",
    "questionary>=1.10.0",
]
 
keywords = ["azure", "container-apps", "cli", "orchestration", "devops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers", 
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Tools",
    "Topic :: System :: Systems Administration",
]

[project.urls]
Homepage = "https://github.com/dsouflis/maestro"
Repository = "https://github.com/dsouflis/maestro" 
Issues = "https://github.com/dsouflis/maestro/issues"

[project.scripts]
maestro = "main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0", 
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
